Rename connectors, cleanup dockerfile and docker compose file
